Refine install builder
