+++ docs/install.md
@@ -5,7 +5,7 @@ title: Install
 
 <script setup>
   import PageHeader from "./.vitepress/theme/components/PageHeader.vue";
-  import InstallScriptGenerator from "./.vitepress/theme/components/InstallScriptGenerator.vue";
+  import InstallBuilder from "./.vitepress/theme/components/InstallBuilder.vue";
   import { options } from "./_data/options";
 </script>
 
@@ -13,5 +13,5 @@ title: Install
   <PageHeader>
     <template #title>Install</template>
   </PageHeader>
-  <InstallScriptGenerator :options="options"></InstallScriptGenerator>
+  <InstallBuilder :options="options"></InstallBuilder>
 </InstallPage>
