+++ package.json
@@ -22,6 +22,7 @@
 		"alpinejs": "3.13.1",
 		"chart.js": "4.4.0",
 		"check-password-strength": "2.0.7",
+		"floating-vue": "^2.0.0-beta.24",
 		"nanoid": "5.0.1",
 		"normalize.css": "8.0.1",
 		"xterm": "5.3.0",
